Доработка сброса пароля версия 3
