+++ src/main/java/com/avito/service/impl/UserServiceImpl.java
@@ -2,18 +2,19 @@ package com.avito.service.impl;
 
 import com.avito.models.User;
 import com.avito.models.posting.Posting;
+import com.avito.models.PasswordResetToken;
+import com.avito.repository.PasswordResetTokenRepository;
 import com.avito.repository.PostingRepository;
 import com.avito.repository.UserRepository;
 import com.avito.service.interfaces.UserService;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.CrossOrigin;
 
+import java.util.Calendar;
 import java.util.List;
 
 @Service
@@ -24,9 +25,9 @@ public class UserServiceImpl implements UserService {
     private static final Logger logger = LoggerFactory.getLogger(UserServiceImpl.class);
 
     private final PasswordEncoder passwordEncoder;
-
     private final UserRepository userRepository;
     private final PostingRepository postingRepository;
+    private final PasswordResetTokenRepository passwordResetTokenRepository;
 
     @Override
     public List<User> getAllUsers() {
@@ -73,5 +74,34 @@ public class UserServiceImpl implements UserService {
         Posting posting = postingRepository.getOne(id_posting);
         user.getFavoritePostings().remove(posting);
         userRepository.save(user);
+    }
+    @Override
+    public void addPasswordToken(PasswordResetToken passwordResetToken) {
+        passwordResetTokenRepository.save(passwordResetToken);
+    }
+    @Override
+    public String validatePasswordResetToken(String token) {
+        PasswordResetToken passToken = passwordResetTokenRepository.findByToken(token);
+        return !isTokenFound(passToken) ? "invalidToken"
+                : isTokenExpired(passToken) ? "expired"
+                : null;
+    }
+    private boolean isTokenFound(PasswordResetToken passToken) {
+        return passToken != null;
+    }
+
+    private boolean isTokenExpired(PasswordResetToken passToken) {
+        final Calendar cal = Calendar.getInstance();
+        return passToken.getExpiryDate().before(cal.getTime());
+    }
+    @Override
+    public void createPasswordResetTokenForUser(User user, String token) {
+        PasswordResetToken myToken = new PasswordResetToken(token, user);
+        addPasswordToken(myToken);
+    }
+
+    @Override
+    public PasswordResetToken findByToken(String token) {
+        return passwordResetTokenRepository.findByToken(token);
     }
 }
