+++ src/main/java/com/avito/service/interfaces/UserService.java
@@ -2,7 +2,7 @@ package com.avito.service.interfaces;
 
 
 import com.avito.models.User;
-import com.avito.models.posting.Posting;
+import com.avito.models.PasswordResetToken;
 
 import java.util.List;
 
@@ -14,4 +14,8 @@ public interface UserService {
         void deleteUser (long id);
         User addFavoritePosting(Long id_posting, Long id_user);
         void deleteFavoritePosting(Long id_posting, Long id_user);
+        void addPasswordToken(PasswordResetToken passwordResetToken);
+        String validatePasswordResetToken(String token);
+        void createPasswordResetTokenForUser(User user, String token);
+        PasswordResetToken findByToken(String token);
 }
