+++ src/main/resources/application.yml
@@ -1,5 +1,6 @@
 server:
   port: 5556
+  domain: localhost
 
 spring:
   datasource:
