+++ src/main/resources/templates/updatePassword.html
@@ -4,7 +4,7 @@
 <body>
 <div><!--sec:authorize="hasAuthority('CHANGE__PASSWORD__PRIVILEGE')"-->
     <h1 th:text="#{message.resetYourPassword}">reset</h1>
-    <form action="/user/savePassword" method="post">
+    <form action="/reset/savePassword" method="post">
         <input name="token" th:value="${token}" type="hidden">
         <label th:text="#{label.user.password}">password</label>
         <input  name="password" type="password" value="" required/>
